Safari-611.1.10/LayoutTests/http/tests/cache/disk-cache/disk-cache-validation-back-navigation-policy-high-priority-expected.txt

Test permutations of various cache headers when using back navigation cache policy (with high priority)

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


running 54 tests

--------Testing loads from disk cache--------
response headers: undefined
response source: Disk cache

response headers: {"Cache-control":"max-age=0"}
response source: Disk cache

response headers: {"Cache-control":"max-age=100"}
response source: Disk cache

response headers: {"Cache-control":"no-store"}
response source: Network

response headers: {"Cache-control":"max-age=0, no-store"}
response source: Network

response headers: {"Cache-control":"max-age=100, no-store"}
response source: Network

response headers: {"Cache-control":"no-cache"}
response source: Disk cache

response headers: {"Cache-control":"max-age=0, no-cache"}
response source: Disk cache

response headers: {"Cache-control":"max-age=100, no-cache"}
response source: Disk cache

response headers: {"ETag":"match"}
response source: Disk cache

response headers: {"Cache-control":"max-age=0","ETag":"match"}
response source: Disk cache

response headers: {"Cache-control":"max-age=100","ETag":"match"}
response source: Disk cache

response headers: {"Cache-control":"no-store","ETag":"match"}
response source: Network

response headers: {"Cache-control":"max-age=0, no-store","ETag":"match"}
response source: Network

response headers: {"Cache-control":"max-age=100, no-store","ETag":"match"}
response source: Network

response headers: {"Cache-control":"no-cache","ETag":"match"}
response source: Disk cache

response headers: {"Cache-control":"max-age=0, no-cache","ETag":"match"}
response source: Disk cache

response headers: {"Cache-control":"max-age=100, no-cache","ETag":"match"}
response source: Disk cache

response headers: {"ETag":"nomatch"}
response source: Disk cache

response headers: {"Cache-control":"max-age=0","ETag":"nomatch"}
response source: Disk cache

response headers: {"Cache-control":"max-age=100","ETag":"nomatch"}
response source: Disk cache

response headers: {"Cache-control":"no-store","ETag":"nomatch"}
response source: Network

response headers: {"Cache-control":"max-age=0, no-store","ETag":"nomatch"}
response source: Network

response headers: {"Cache-control":"max-age=100, no-store","ETag":"nomatch"}
response source: Network

response headers: {"Cache-control":"no-cache","ETag":"nomatch"}
response source: Disk cache

response headers: {"Cache-control":"max-age=0, no-cache","ETag":"nomatch"}
response source: Disk cache

response headers: {"Cache-control":"max-age=100, no-cache","ETag":"nomatch"}
response source: Disk cache

response headers: {"Cache-control":"must-revalidate"}
response source: Disk cache

response headers: {"Cache-control":"max-age=0, must-revalidate"}
response source: Disk cache

response headers: {"Cache-control":"max-age=100, must-revalidate"}
response source: Disk cache

response headers: {"Cache-control":"no-store, must-revalidate"}
response source: Network

response headers: {"Cache-control":"max-age=0, no-store, must-revalidate"}
response source: Network

response headers: {"Cache-control":"max-age=100, no-store, must-revalidate"}
response source: Network

response headers: {"Cache-control":"no-cache, must-revalidate"}
response source: Disk cache

response headers: {"Cache-control":"max-age=0, no-cache, must-revalidate"}
response source: Disk cache

response headers: {"Cache-control":"max-age=100, no-cache, must-revalidate"}
response source: Disk cache

response headers: {"ETag":"match","Cache-control":"must-revalidate"}
response source: Disk cache

response headers: {"Cache-control":"max-age=0, must-revalidate","ETag":"match"}
response source: Disk cache

response headers: {"Cache-control":"max-age=100, must-revalidate","ETag":"match"}
response source: Disk cache

response headers: {"Cache-control":"no-store, must-revalidate","ETag":"match"}
response source: Network

response headers: {"Cache-control":"max-age=0, no-store, must-revalidate","ETag":"match"}
response source: Network

response headers: {"Cache-control":"max-age=100, no-store, must-revalidate","ETag":"match"}
response source: Network

response headers: {"Cache-control":"no-cache, must-revalidate","ETag":"match"}
response source: Disk cache

response headers: {"Cache-control":"max-age=0, no-cache, must-revalidate","ETag":"match"}
response source: Disk cache

response headers: {"Cache-control":"max-age=100, no-cache, must-revalidate","ETag":"match"}
response source: Disk cache

response headers: {"ETag":"nomatch","Cache-control":"must-revalidate"}
response source: Disk cache

response headers: {"Cache-control":"max-age=0, must-revalidate","ETag":"nomatch"}
response source: Disk cache

response headers: {"Cache-control":"max-age=100, must-revalidate","ETag":"nomatch"}
response source: Disk cache

response headers: {"Cache-control":"no-store, must-revalidate","ETag":"nomatch"}
response source: Network

response headers: {"Cache-control":"max-age=0, no-store, must-revalidate","ETag":"nomatch"}
response source: Network

response headers: {"Cache-control":"max-age=100, no-store, must-revalidate","ETag":"nomatch"}
response source: Network

response headers: {"Cache-control":"no-cache, must-revalidate","ETag":"nomatch"}
response source: Disk cache

response headers: {"Cache-control":"max-age=0, no-cache, must-revalidate","ETag":"nomatch"}
response source: Disk cache

response headers: {"Cache-control":"max-age=100, no-cache, must-revalidate","ETag":"nomatch"}
response source: Disk cache

--------Testing loads through memory cache (XHR behavior)--------
response headers: undefined
response source: Disk cache

response headers: {"Cache-control":"max-age=0"}
response source: Disk cache

response headers: {"Cache-control":"max-age=100"}
response source: Memory cache

response headers: {"Cache-control":"no-store"}
response source: Network

response headers: {"Cache-control":"max-age=0, no-store"}
response source: Network

response headers: {"Cache-control":"max-age=100, no-store"}
response source: Network

response headers: {"Cache-control":"no-cache"}
response source: Disk cache

response headers: {"Cache-control":"max-age=0, no-cache"}
response source: Disk cache

response headers: {"Cache-control":"max-age=100, no-cache"}
response source: Disk cache

response headers: {"ETag":"match"}
response source: Memory cache after validation

response headers: {"Cache-control":"max-age=0","ETag":"match"}
response source: Memory cache after validation

response headers: {"Cache-control":"max-age=100","ETag":"match"}
response source: Memory cache

response headers: {"Cache-control":"no-store","ETag":"match"}
response source: Network

response headers: {"Cache-control":"max-age=0, no-store","ETag":"match"}
response source: Network

response headers: {"Cache-control":"max-age=100, no-store","ETag":"match"}
response source: Network

response headers: {"Cache-control":"no-cache","ETag":"match"}
response source: Memory cache after validation

response headers: {"Cache-control":"max-age=0, no-cache","ETag":"match"}
response source: Memory cache after validation

response headers: {"Cache-control":"max-age=100, no-cache","ETag":"match"}
response source: Memory cache after validation

response headers: {"ETag":"nomatch"}
response source: Network

response headers: {"Cache-control":"max-age=0","ETag":"nomatch"}
response source: Network

response headers: {"Cache-control":"max-age=100","ETag":"nomatch"}
response source: Memory cache

response headers: {"Cache-control":"no-store","ETag":"nomatch"}
response source: Network

response headers: {"Cache-control":"max-age=0, no-store","ETag":"nomatch"}
response source: Network

response headers: {"Cache-control":"max-age=100, no-store","ETag":"nomatch"}
response source: Network

response headers: {"Cache-control":"no-cache","ETag":"nomatch"}
response source: Network

response headers: {"Cache-control":"max-age=0, no-cache","ETag":"nomatch"}
response source: Network

response headers: {"Cache-control":"max-age=100, no-cache","ETag":"nomatch"}
response source: Network

response headers: {"Cache-control":"must-revalidate"}
response source: Disk cache

response headers: {"Cache-control":"max-age=0, must-revalidate"}
response source: Disk cache

response headers: {"Cache-control":"max-age=100, must-revalidate"}
response source: Memory cache

response headers: {"Cache-control":"no-store, must-revalidate"}
response source: Network

response headers: {"Cache-control":"max-age=0, no-store, must-revalidate"}
response source: Network

response headers: {"Cache-control":"max-age=100, no-store, must-revalidate"}
response source: Network

response headers: {"Cache-control":"no-cache, must-revalidate"}
response source: Disk cache

response headers: {"Cache-control":"max-age=0, no-cache, must-revalidate"}
response source: Disk cache

response headers: {"Cache-control":"max-age=100, no-cache, must-revalidate"}
response source: Disk cache

response headers: {"ETag":"match","Cache-control":"must-revalidate"}
response source: Memory cache after validation

response headers: {"Cache-control":"max-age=0, must-revalidate","ETag":"match"}
response source: Memory cache after validation

response headers: {"Cache-control":"max-age=100, must-revalidate","ETag":"match"}
response source: Memory cache

response headers: {"Cache-control":"no-store, must-revalidate","ETag":"match"}
response source: Network

response headers: {"Cache-control":"max-age=0, no-store, must-revalidate","ETag":"match"}
response source: Network

response headers: {"Cache-control":"max-age=100, no-store, must-revalidate","ETag":"match"}
response source: Network

response headers: {"Cache-control":"no-cache, must-revalidate","ETag":"match"}
response source: Memory cache after validation

response headers: {"Cache-control":"max-age=0, no-cache, must-revalidate","ETag":"match"}
response source: Memory cache after validation

response headers: {"Cache-control":"max-age=100, no-cache, must-revalidate","ETag":"match"}
response source: Memory cache after validation

response headers: {"ETag":"nomatch","Cache-control":"must-revalidate"}
response source: Network

response headers: {"Cache-control":"max-age=0, must-revalidate","ETag":"nomatch"}
response source: Network

response headers: {"Cache-control":"max-age=100, must-revalidate","ETag":"nomatch"}
response source: Memory cache

response headers: {"Cache-control":"no-store, must-revalidate","ETag":"nomatch"}
response source: Network

response headers: {"Cache-control":"max-age=0, no-store, must-revalidate","ETag":"nomatch"}
response source: Network

response headers: {"Cache-control":"max-age=100, no-store, must-revalidate","ETag":"nomatch"}
response source: Network

response headers: {"Cache-control":"no-cache, must-revalidate","ETag":"nomatch"}
response source: Network

response headers: {"Cache-control":"max-age=0, no-cache, must-revalidate","ETag":"nomatch"}
response source: Network

response headers: {"Cache-control":"max-age=100, no-cache, must-revalidate","ETag":"nomatch"}
response source: Network

--------Testing loads through memory cache (subresource behavior)--------
response headers: undefined
response source: Disk cache

response headers: {"Cache-control":"max-age=0"}
response source: Disk cache

response headers: {"Cache-control":"max-age=100"}
response source: Memory cache

response headers: {"Cache-control":"no-store"}
response source: Network

response headers: {"Cache-control":"max-age=0, no-store"}
response source: Network

response headers: {"Cache-control":"max-age=100, no-store"}
response source: Network

response headers: {"Cache-control":"no-cache"}
response source: Disk cache

response headers: {"Cache-control":"max-age=0, no-cache"}
response source: Disk cache

response headers: {"Cache-control":"max-age=100, no-cache"}
response source: Disk cache

response headers: {"ETag":"match"}
response source: Memory cache after validation

response headers: {"Cache-control":"max-age=0","ETag":"match"}
response source: Memory cache after validation

response headers: {"Cache-control":"max-age=100","ETag":"match"}
response source: Memory cache

response headers: {"Cache-control":"no-store","ETag":"match"}
response source: Network

response headers: {"Cache-control":"max-age=0, no-store","ETag":"match"}
response source: Network

response headers: {"Cache-control":"max-age=100, no-store","ETag":"match"}
response source: Network

response headers: {"Cache-control":"no-cache","ETag":"match"}
response source: Memory cache after validation

response headers: {"Cache-control":"max-age=0, no-cache","ETag":"match"}
response source: Memory cache after validation

response headers: {"Cache-control":"max-age=100, no-cache","ETag":"match"}
response source: Memory cache after validation

response headers: {"ETag":"nomatch"}
response source: Network

response headers: {"Cache-control":"max-age=0","ETag":"nomatch"}
response source: Network

response headers: {"Cache-control":"max-age=100","ETag":"nomatch"}
response source: Memory cache

response headers: {"Cache-control":"no-store","ETag":"nomatch"}
response source: Network

response headers: {"Cache-control":"max-age=0, no-store","ETag":"nomatch"}
response source: Network

response headers: {"Cache-control":"max-age=100, no-store","ETag":"nomatch"}
response source: Network

response headers: {"Cache-control":"no-cache","ETag":"nomatch"}
response source: Network

response headers: {"Cache-control":"max-age=0, no-cache","ETag":"nomatch"}
response source: Network

response headers: {"Cache-control":"max-age=100, no-cache","ETag":"nomatch"}
response source: Network

response headers: {"Cache-control":"must-revalidate"}
response source: Disk cache

response headers: {"Cache-control":"max-age=0, must-revalidate"}
response source: Disk cache

response headers: {"Cache-control":"max-age=100, must-revalidate"}
response source: Memory cache

response headers: {"Cache-control":"no-store, must-revalidate"}
response source: Network

response headers: {"Cache-control":"max-age=0, no-store, must-revalidate"}
response source: Network

response headers: {"Cache-control":"max-age=100, no-store, must-revalidate"}
response source: Network

response headers: {"Cache-control":"no-cache, must-revalidate"}
response source: Disk cache

response headers: {"Cache-control":"max-age=0, no-cache, must-revalidate"}
response source: Disk cache

response headers: {"Cache-control":"max-age=100, no-cache, must-revalidate"}
response source: Disk cache

response headers: {"ETag":"match","Cache-control":"must-revalidate"}
response source: Memory cache after validation

response headers: {"Cache-control":"max-age=0, must-revalidate","ETag":"match"}
response source: Memory cache after validation

response headers: {"Cache-control":"max-age=100, must-revalidate","ETag":"match"}
response source: Memory cache

response headers: {"Cache-control":"no-store, must-revalidate","ETag":"match"}
response source: Network

response headers: {"Cache-control":"max-age=0, no-store, must-revalidate","ETag":"match"}
response source: Network

response headers: {"Cache-control":"max-age=100, no-store, must-revalidate","ETag":"match"}
response source: Network

response headers: {"Cache-control":"no-cache, must-revalidate","ETag":"match"}
response source: Memory cache after validation

response headers: {"Cache-control":"max-age=0, no-cache, must-revalidate","ETag":"match"}
response source: Memory cache after validation

response headers: {"Cache-control":"max-age=100, no-cache, must-revalidate","ETag":"match"}
response source: Memory cache after validation

response headers: {"ETag":"nomatch","Cache-control":"must-revalidate"}
response source: Network

response headers: {"Cache-control":"max-age=0, must-revalidate","ETag":"nomatch"}
response source: Network

response headers: {"Cache-control":"max-age=100, must-revalidate","ETag":"nomatch"}
response source: Memory cache

response headers: {"Cache-control":"no-store, must-revalidate","ETag":"nomatch"}
response source: Network

response headers: {"Cache-control":"max-age=0, no-store, must-revalidate","ETag":"nomatch"}
response source: Network

response headers: {"Cache-control":"max-age=100, no-store, must-revalidate","ETag":"nomatch"}
response source: Network

response headers: {"Cache-control":"no-cache, must-revalidate","ETag":"nomatch"}
response source: Network

response headers: {"Cache-control":"max-age=0, no-cache, must-revalidate","ETag":"nomatch"}
response source: Network

response headers: {"Cache-control":"max-age=100, no-cache, must-revalidate","ETag":"nomatch"}
response source: Network

PASS successfullyParsed is true

TEST COMPLETE
